Добавил логику для появления комментариев
