+++ js/main.js
@@ -1,3 +1,5 @@
 import { createArrayPhoto } from './data.js';
 import { miniaturesList } from './miniatures.js';
 import { miniaturePictures } from './modal-big-picture.js';
+import { logic1 } from './list-comment-logic.js';
+import { logic2 } from './list-comment-logic.js';
