+++ js/modal-big-picture.js
@@ -1,6 +1,8 @@
 import { miniaturesList } from './miniatures.js';
 import { getCommentsList } from './comments-list.js';
 import { isEscapeKey } from './util.js';
+import { logic1 } from './list-comment-logic.js';
+import { logic2 } from './list-comment-logic.js';
 
 
 const modalBigPicture = document.querySelector('.big-picture');
@@ -43,38 +45,61 @@ miniaturePictures.forEach((miniaturePicture) => {
     infoModalBigPicture.querySelector('.social__comment-total-count').textContent = miniaturePicture.querySelector('.picture__comments').textContent;
     const currentId = miniaturePicture.querySelector('.picture__img').id;
     getCommentsList(currentId);
+    logic2();
+
+    // document.querySelector('.social__comment-shown-count').textContent = COMMENTS_UPLOAD_VOLUME;
+    // const socialComment = document.querySelectorAll('.social__comment');
+    // socialComment.forEach((element) => element.classList.add('hidden'));
+
+    // if (socialComment.length <= COMMENTS_UPLOAD_VOLUME) {
+    //   showButton.classList.add('hidden');
+    // }
+
+    // let current = 5;
+
+    // const proverka = (increment) => {
+
+    //   for (let i = 0 + increment; i < current && current; i++) {
+    //     if (current > socialComment.length) {
+    //       current = socialComment.length;
+    //     }
+    //     socialComment[i].classList.remove('hidden');
+
+    //   } current = current + COMMENTS_UPLOAD_VOLUME;
+    // };
+    // proverka(0);
+
+    // showButton.addEventListener('click', () => {
+    //   const increment = COMMENTS_UPLOAD_VOLUME;
+    //   proverka(increment);
+    // });
 
 
-    document.querySelector('.social__comment-shown-count').textContent = COMMENTS_UPLOAD_VOLUME;
-    const socialComment = document.querySelectorAll('.social__comment');
-    if (socialComment.length <= COMMENTS_UPLOAD_VOLUME) {
-      showButton.classList.add('hidden');
-    }
-
-    for (let i = COMMENTS_UPLOAD_VOLUME; i < socialComment.length; i++) {
-      socialComment[i].classList.add('hidden');
-    }
   });
+
 });
 
 closeModalButton.addEventListener('click', () => {
   closeModalBigPicture();
 });
 
+logic1();
+
 export { miniaturePictures };
+export { showButton };
+export { COMMENTS_UPLOAD_VOLUME };
 
 
-showButton.addEventListener('click', () => {
-  const showCount = document.querySelector('.social__comment-shown-count');
-  const socialComment = document.querySelectorAll('.social__comment');
-  let showCountValue = +showCount.textContent;
+// showButton.addEventListener('click', () => {
+//   const showCount = document.querySelector('.social__comment-shown-count');
+//   const socialComment = document.querySelectorAll('.social__comment');
+//   let showCountValue = +showCount.textContent;
 
-  if (showCountValue <= socialComment.length && socialComment.length - showCountValue > COMMENTS_UPLOAD_VOLUME) {
-    showCount.textContent = showCountValue + COMMENTS_UPLOAD_VOLUME;
-
-  } else {
-    showCount.textContent = socialComment.length - showCountValue + showCountValue;
-    showButton.classList.add('hidden');
-    showCountValue = 0;
-  }
-});
+//   if (showCountValue <= socialComment.length && socialComment.length - showCountValue > COMMENTS_UPLOAD_VOLUME) {
+//     showCount.textContent = showCountValue + COMMENTS_UPLOAD_VOLUME;
+//   } else {
+//     showCount.textContent = socialComment.length - showCountValue + showCountValue;
+//     showButton.classList.add('hidden');
+//     showCountValue = 0;
+//   }
+// });
